Assign resource limits to Elastic Components
Set resource limits for Standard configuration and AIO,
to ensure sufficient resources allocated. This also ensures there
are no "BestEffort" (without limits set) elastic component pods.

These resource limits are also introduced within the helm-overrides,
so the resources may be updated and applied as required.
Further system setting of resource limits based upon available
controller resources is to be considered in subsequent commit.

Update host-label to allow the assignment/removal of host-labels which
do not require host-lock/unlock cycle while unlocked, as deployment is
more efficient as it does not require another cycle of the
host to be locked.

Change-Id: I05659c2dc8ec4211250e37a73e9d1791fefa59eb
Story: 2005733
Task: 37160
Depends-On: https://review.opendev.org/688996
Signed-off-by: John Kung <john.kung@windriver.com>