Run sysinv-api as sysinv user and normal groups
The sysinv-api is currently running as the sysinv user, but with
the group set to root. This is not correct - the sysinv-api should
run with the normal groups for the sysinv user, which include
sys_protected.

This change is necessary to allow the sysinv-api to communicate
with the kubernetes API (the configuration file is using group
sys_protected).

Change-Id: I85edd4a10b3aa1d537fc6c47af411bf69b18516b
Story: 2006590
Task: 36724
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>