Copy sysinv.config to drbd drive only once
This change is to ensure copying sysinv.config to drbd drive only once
when the controller-0 is unlocked successfully the 1st time.

The file was previously copied periodically from the active controller
which causes occasionally fail to apply the runtime config if a swact
happens to start. The result is that the controller is config-out-date.

This fix also ensure all nodes going through the same invprovision
cycle: none (installing)
       -> provisioning (installed)
       -> provisioned (unlocked successfully 1st time)
Previously the controller-0 of a standard load starts with provisioned,
even before it is unlocked the 1st time.

Depends-on: https://review.opendev.org/691713
Closes-Bug: 1848355
Closes-Bug: 1849837

Change-Id: Ifdfb985c8cde86f9c9074dc0774cb0dd805aacd1
Signed-off-by: Bin Qian <bin.qian@windriver.com>