Fix tox in other repos that import sysinv
sysinv.common.utils is imported in cgcs-patch.

When keyring was added to that file, the tox and
zuul jobs in cgcs-patch failed.

The root reason for this is that when those other repos
add sysinv as a dependency, only the requirements.txt is
used when constructing the egg for import by tox.

By properly relocating runtime required files from
test-requirements.txt to requirements.txt, those other
repos are able to import the additional required components.

This does not impact rpms since those are driven by the
Requires directives in the spec files.

Change-Id: I8d9d84345a0810e311d2a50cdcbdb49056f80c2c
Fixes-Bug: 1848786
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>