Run sysinv-api as sysinv user and normal groups
The sysinv-api is currently running as the sysinv user, but with
the group set to root. This is not correct - the sysinv-api should
run with the normal groups for the sysinv user, which include
sys_protected.

This change is necessary to allow the sysinv-api to communicate
with the kubernetes API (the configuration file is using group
sys_protected).

Change-Id: I0ebd60b256cf337495079fdb723b3ef2836e56de
Story: 2006590
Task: 36724
Depends-On: https://review.opendev.org/690418
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>