Metric set event and rate tuning
Adjust the set of metrics that are generated by filtering out events that are
not required.  In addition, the sets of metrics have been adjusted to reflect
the proper scoping to host vs pod resoures.  Finally, the rates have been
adjusted to align host and state metrics to 60s, the kubernetes resource
metrics to 10s and the file system metrics have been extended to 5mins.

Story: 2005733
Task: 37275
Change-Id: I5d03f9c24a257b61165a59a8329a8cc2156132e0
Signed-off-by: Matt Peters <matt.peters@windriver.com>