Set uniform PM-QoS value for all cpus on lowlatency nodes
We observed poor cyclictest results on lowlatency nodes. This was
due to the fact that we were trying to save power by only marking
in-use CPUs as low-latency, but there was an underlying issue where
any high-latency CPUs result in a latency hit for all other CPUs.

As a workaround to provide consistent low-latency performance, set
all CPUs to only sleep as deep as C1 on lowlatency nodes.

Change-Id: Idd627eee6e7ff95939e428b180d07d8ab1cadad2
Partial-Bug: 1845735
Signed-off-by: Jim Gauld <james.gauld@windriver.com>