Support to preserve user overrides when updating app to a new version
This commit updates to provide the ability to preserve the user
overrides from the old version when updating app to a new version.

The key "maintain_user_overrides" in application metadata file
indicates if the user overrides will be reused or not during
update, user can specify --reuse-user-overrides <true/false>
to override the metadata preference specified by the application.

This commit also updates armada log for each application operation
to include timestamp in its log name and keep the maximum 10 logs
for each type of log.

Change-Id: I1c077017c8a60dc63d27135ff5dd966aa38387f8
Partial-Bug: 1849486
Signed-off-by: Angie Wang <angie.wang@windriver.com>