Create sysinv-utils for generating host-overrides
This commits creates a new utility command /usr/bin/sysinv-utils
that can be used by ansible playbooks to generate the host overrides
file that will be used for upgrades and backup/restore.

This commits also enhanced the following CLI commands to include a
--column and --format options for better CLI outputs data manipulation.
	host-show
	host-list
	service-parameter-list
        application-show

These two new options are optional. The default behavior is not changed.

Story: 2006590
Task: 36892
Change-Id: I36b8ea411a4eb74d6e1a2194982025905e9299e5
Signed-off-by: Kristine Bujold <kristine.bujold@windriver.com>