Update list of skipped bandit plugins.
Adding a zuul job so that any new bandit failures
will be caught.

Switching bandit to run in python3.

B111 plugin was removed from bandit by:
PyCQA/bandit@d93eed5

Cleanup the list of bandit checks that need to be skipped
in order for this tox job to pass. Those checks can be
removed when the code is updated by future commits.

Story: 2004515
Task: 37744
Change-Id: I1809b1124a5a693d288fb7ea893f30c24487af36
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>