Allow host-delete for unprovisioned hosts without mtce notify
In the case of host-delete in which hostname or personality has not yet
been provisioned, host-delete must not send the notification to mtce.
This is because mtce is not notified by sysinv about hosts which have
not yet been provisioned with a hostname or personality.

Change-Id: Iec310641411520ab2e1b711632c6c64e251ac64b
Closes-Bug: 1850683
Signed-off-by: John Kung <john.kung@windriver.com>
(cherry picked from commit 64e097037a1c0cd1aedcb3d1220f4f48ffb379f4)