Allow configuration of PTP master/slave interfaces
With this change a new field ptp_role will be added to the interfaces
table. This will allow administrators to configure which interfaces will
be used for PTP serivces. Only platform or SRIOV interfaces may be
specified for PTP use. If a host has clock_synchronization=ptp there
must be at least one host interface with a ptp_role value specified.

Change-Id: I86eddebf2f716f76d4eb9a5b5698a61ff3aec566
Story: 2006759
Task: 37690
Signed-off-by: David Sullivan <david.sullivan@windriver.com>