Create default Monitor config for Distributed Cloud
In order to automate the required steps in Distributed Cloud
configuration for stx-monitor, this configures the Armada charts
and system helm overrides for the subcloud to send to
the Elasticsearch cluster on the SystemController.

As the host-label semantic checks are based upon the Armada
charts enabled, the host-label required on apply are
also correctly adjusted.

With this update, no additional configuration steps are
required in the sublcloud to index into the SystemController,
other than the standard 'application-upload', 'application-apply'.

Verified on Distributed Cloud System, Standard and AIO.

Change-Id: I5767553a3bf52afe4790f94ac4fbdd38454a35c6
Story: 2006990
Task: 37755
Signed-off-by: John Kung <john.kung@windriver.com>