B&R: Fix restore of stx-openstack
In the first part of the application restore, the 'system apply'
command is used with the 'mode' argument.
When the sysinv checks for changes in the overrides, it doesn't
take into account the 'mode' argument. This leads
to an incorrect reapply of the app.

The solution is to save the 'mode' argument in the sysinv
database. This is now used when evaluating the app overrides
for changes, also when reapplying the app.

Story: 2006770
Task: 37780
Change-Id: I73f5a83bb1dd3a9e1fed021ab74af53463012189
Signed-off-by: Mihnea Saracin <Mihnea.Saracin@windriver.com>