B&R: Fix openstack helper scripts for B&R
The helper script will be used to backup and
restore the glance images. This script was previously
deleted by https://review.opendev.org/#/c/631312/.
In order to be able to access the containerized openstack,
the tidy storage script was modified to use the configuration
data from the '/etc/openstack/clouds.yaml' file.
Also, the numpy dependency was removed because all
the numpy code was easily rewritten in plain python code.
Moreover, the v1 glance client was removed and the code
was updated accordingly.

Depends-on: https://review.opendev.org/#/c/698478/
Story: 2006770
Task: 37286
Change-Id: I2918b4220b1bd3aafde20263e6cef756ac06c6db
Signed-off-by: Mihnea Saracin <Mihnea.Saracin@windriver.com>