Import subprocess from eventlet.green package
The python kubernetes client requires a newer version of eventlet.
However, newer versions of eventlet have an issue with the subprocess
module, which requires subprocess to be imported from eventlet.green
instead of being imported directly. See:
https://github.com/eventlet/eventlet/issues/413

To allow eventlet to be upversioned (from 0.18.4 to 0.24.1), the
subprocess imports are being changed.

Change-Id: I884e8e539864f26b3593e9c15a29bbd099c94453
Story: 2006781
Task: 37583
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>