Skip generating hieradata for uninventoried host
When _config_apply_runtime_manifest is run with force=True, it will
generate hieradata for all hosts, even hosts that have yet to come
online and are still uninventoried.  Because the node is still
uninventoried, the hieradata will have an incomplete network config
with only the loopback interface configured.
When those hosts come online for the first time, they will mount
/opt/platform and check for their hieradata, and because it exists,
apply their puppet manifests with that hieradata, which will bring down
all non-loopback interfaces and make the node go offline and become
unreachable.
This commit adds a check to only generate hieradata for inventoried
hosts, even when force=True.

Closes-bug: 1856617
Change-Id: I37c611e9a1047d5651844c71351f40d939977b6f
Signed-off-by: Joseph Richard <joseph.richard@windriver.com>