Move incorrectly located stand-alone scripts
Moving incorrectly located stand-alone scripts from the
sysinv/cmd directory to the scripts directory. These scripts
appeared to be using eventlet due to the __init__.py file in the
sysinv/cmd directory, but were not, as they are moved to the
/bin directory at installation time.

This is being done in preparation for upversioning eventlet.

Change-Id: I511d61e14b3eb1511cee03aa0e36ffd1277e79ab
Story: 2006781
Task: 37583
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>