Remove sqlalchemy-migrate 0.7.3 patching
This is similar to upstream submissions in 2013 such as
https://review.opendev.org/#/c/56705/

We currently build, test and ship with a newer version
of sqlalchemy-migrate 0.11.0 so we do not need to handle
0.7.3 workarounds and patches.

This review also fixes: tests.db.sqlalchemy.test_migrations
which would fail when tested in isolation, due to its environment
not being initialized properly, but working if certain other tests
were run first.

Some skipped tests are now also enabled.

Change-Id: I6ce73491f7a2471f25c7691dab8b53dc354619a2
Story: 2004515
Task: 37954
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>