stx-monitor app: enhanced missing elastic label error message
When attempting to apply stx-monitor app without the required elastic
labels, user will get an error message detailing what is missing.
This label was enhanced to be more readable and provide instructions
on how to apply the missing labels.

Moved utility methods up to the super class LabelTestCase. Also
moved host definitions to each of the test classes since
LabelAssignTestCase require hosts to be locked and
StxAppLabelsTest require hosts to be unlocked.

Added the following enhanced logic:
1. All controller nodes are required to have all 4 elastic labels
2. At least one worker node should have elastic-master label.
3. Add label value validation and print error message for invalid
values.

Change-Id: I0b2bc7b87d2e8bf78bfbba455d499c52e1a72349
Closes-Bug: 1856763
Signed-off-by: abdelfattah saafan <abdelfattah.saafan@windriver.com>