Provide the host UUID when updating a route on a host interface
Optimize updating routes on a host's network interface by sending the
host UUID with the config data when requesting the route to be updated.

Without providing the host UUID all personalities will receive the route
update request and apply a manifest regardless of if the interface is
actually present on the host.

With the host UUID provided the sysinv agent can skip the puppet apply
if the host's UUID doesn't match.

Change-Id: I3ed0fd657a1d0a4fafd4ff15c3b81da0cb8503e0
Story: 2007101
Task: 38145
Signed-off-by: Robert Church <robert.church@windriver.com>