Workaround for cleaning up MatchNodeSelector pods after host reboot
Added a K8sPodOperator class to look for and remove Failed pods with a
MatchNodeSelector reason.

MatchNodeSelector pods related to applications will not be removed by
K8S automatically. These pods may block subsequent application applies
as tiller expects these pods to be in a non failed state.

A check for this condition is added in two locations:
- to the _k8s_application_audit() which is run immediately on
  sysinv-conductor startup and runs every minute. This runs 4 times in a
  5 minute window at startup on a simplex install. This should catch all
  cases unless there is a delay accessing the k8s API that lasts longer
  than 5 minutes at startup.
- to the application-apply path. This would cover any case that occurs
  after the initial 5 minute conductor startup OR any occurance on a
  non-simplex installation (so far only observed on AIO-SX)

NOTE: This commit will be reverted once a proper upstream k8S fix is
provided.

Related upstream bugs:
- https://github.com/kubernetes/kubernetes/issues/80745
- https://github.com/kubernetes/kubernetes/issues/85334

The following PR was tested and fixed this issue but has not landed
upstream in a new k8s release:
- https://github.com/kubernetes/kubernetes/pull/80976

Change-Id: Ia5418794a44e7821933e8335d5c5db25b58a739f
Closes-Bug: #1849688
Signed-off-by: Robert Church <robert.church@windriver.com>