Disable keystone caching on subclouds
The use of keystone caching on subclouds causes problems because
the syncing of fernet keys to the subcloud results in stale
cache entries. This causes authentication failures until the
cache entries age out or new tokens are created.

Since the keystone load in a subcloud is light, there is really
no need for caching at this time - it is being disabled in
subclouds.

Change-Id: I777c57c46cf1bcd701fbbac73228a2cb81d8424b
Closes-Bug: 1860372
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>