Allow users to override a single image with a custom registry
In the case that the user overrides a single image with a
custom registry that is not from any known registries
in Sysinv. This image downloading will fail as it
prepends the docker.io registry to the image reference
, then generates an invalid image tag.

The original purpose of adding that logic is to handle
the image that comes from docker.io but do not have
docker.io explicitly specified in its image name. This
case has already been updated to handle in the class
"AppImageParser".

This commit removes the related logic that causing the
issue.

Tested:
 - system helm-override-update stx-openstack nova openstack \
     --set images.tags.nova_api=mycustomregistry.com/stx-nova:latest
 - system application-apply stx-openstack

Change-Id: I07d1a658c3cf56a3e09e81e1f947f93de50b513d
Closes-Bug: 1859881
Signed-off-by: Angie Wang <angie.wang@windriver.com>