application-apply error string too long
During application-apply exception handling, str(e) is
used as the input to the progress column of the kube_app
table in the database, which may be longer than the 255
character limit.  The result is an application stuck
in 'applying' status.  This update adds a more readable
error message to just check logs.

There are other instances where str(e) is used as input to
the database and could cause a similar problem which should
also be looked at.

Change-Id: I01a5e8f56a628726163e2cfffc58143ae8d5f845
Closes-Bug: 1867019
Signed-off-by: Kevin Smith <kevin.smith@windriver.com>