Fix SR-IOV runtime manifest apply
When an SR-IOV interface is configured, the platform's
network runtime manifest is applied in order to apply the virtual
function (VF) config and restart the interface.  This results in
sysinv being able to determine and populate the puppet hieradata
with the virtual function PCI addresses.

A side effect of the network manifest apply is that potentially
all platform interfaces may be brought down/up if it is determined
that their configuration has changed.  This will likely be the case
for a system which configures SR-IOV interfaces before initial
unlock.

A few issues have been encountered because of this, with some
services not behaving well when the interface they are communicating
over suddenly goes down.

This commit makes the SR-IOV VF configuration much more targeted
so that only the operation of setting the desired number of VFs
is performed.

Closes-Bug: #1868584
Depends-On: https://review.opendev.org/715669
Change-Id: Ie162380d3732eb1b6e9c553362fe68cbc313ae2b
Signed-off-by: Steven Webster <steven.webster@windriver.com>