Fix beat fails to parse kubernetes.pod.labels.app
Metricbeat and Filebeat fail to parse labels that are used as both
objects and keywords in the Elasticsearch document hierarchy. This
change addresses this issue by enabling the 'labels.dedot' and
'annotations.dedot' options on Metricbeat kubernetes modules and
Filebeat's kubernetes metadata processor, which automatically escapes
conflicting labels and annotations.

Story: 2007221
Task: 39463

Change-Id: Id7f6cd6fc499ea4644e16c80b68ebde19c6f59ad
Signed-off-by: Simon Cousineau <Simon.Cousineau@windriver.com>