Update MatchNodeSelector recovery logic for NodeAffinity status
NodeAffinity pods related to applications will not be removed by
K8S automatically. These pods may block subsequent application applies
as tiller expects these pods to be in a non failed state.

This update now will look for NodeAffinity pods when the sysinv
conductor starts. This is no longer limited to simplex nodes. This
behavior is now observed on simplex and duplex controller configurations
as of the upversion to k8s v1.18.1.

Change-Id: I6384ffd1d14ac105e26b83c02aaa8f090e1fdde1
Story: 2006999
Task: 39475
Related-Bug: #1849688
Signed-off-by: Robert Church <robert.church@windriver.com>