Correct alarm calculations in health check
The health check would incorrectly report all alarms as management
affecting. This was a result of moving to the fm API instead of directly
querying the database. As we are querying the API, a tuple is never
returned and the mgmt_affecting property is calculated to "True" or
"False".

Same root cause as this bug/change:
https://review.opendev.org/#/c/664274/

Change-Id: Ia0b8a1df9526daa5052bf977f2c8812416b7e3b9
Story: 2007403
Task: 39517
Signed-off-by: David Sullivan <david.sullivan@windriver.com>