Enable --reserved-cpus option in k8s v1.18.1
The option was introduced in k8s v1.17 and will now be used to define
the explicit set of CPUs that are reserved for specific cpu functions in
StarlingX.

This retires setting the number of CPUs reserved in the --kube-reserved
and --system-reserved options.

Instead of calculating the number of CPUs related to reservations,
provide the specific list of CPUs in a comma separated range format.
This will be used by puppet to set the --reserved-cpus option based on
cpu manager policy.

Remove restrictions around CPU assignments:
- Allow platform cores to be reserved on any processor
- Allow application isolated cores to be reserved on any processor

Change-Id: I1a3d4e4cca7b6940682a787c2e7348e56a047a06
Depends-On: https://review.opendev.org/#/c/722189
Story: 2006999
Task: 39528
Signed-off-by: Robert Church <robert.church@windriver.com>