Fix application-update reuse-user-overrides
The 'maintain_user_overrides' flag in the application tarball
metadata.yaml file is meant to indicate whether to preserve
user overrides over application update.  The --reuse-user-overrides
flag of the application-update command can override the setting
in the metadata.yaml file, but the current logic means the
'maintain_user_overrides' flag will never be checked even if the
--reuse-user-overrides flag is not set.

This update allows the maintain_user_overrides to be checked when
the --reuse-user-overrides flag is not set.

Closes-Bug: 1874552
Change-Id: I38e009f72c432f43b1ad8744771ce32de1269736
Signed-off-by: Kevin Smith <kevin.smith@windriver.com>