Update Logstash to use NodePort Ingress
Logstash should not be using a custom port for collectd
input from the K8s NodePort range since it might cause a
conflict if the port is allocated to another service.
Therefore, logstash will use a proper NodePort value
reserved by the nginx-ingress service.

Do not disable the nginx-ingress on the subcloud since it
is required for collectd to send events to logstash.

Story: 2007221
Task: 39549

Change-Id: Ibdcbcf1b217ddd17197c0e8fb6cc069a573d10a5
Depends-On: https://review.opendev.org/#/c/722674
Signed-off-by: Matt Peters <matt.peters@windriver.com>