Remove storage class backend from helm overrides
Remove the storageClass parameter from the stx-monitor
helm system overrides.  With support for different
storage classes, the specific request for the storage
class of "general" should not be configured so that the
default storage class is used when not specified.

NOTE: The old parameter had no effect since it should
have been storageClassName.  However, it is being removed
since it is confusing to the end user.

Story: 2007391
Task: 39589

Change-Id: Ie690e53404df183337b563902d3566285db27313
Signed-off-by: Matt Peters <matt.peters@windriver.com>