Provide an update strategy for Tiller deployment
In the case of a simplex controller configuration the current patching
strategy for the Tiller environment will fail as the tiller ports will
be in use when the new deployment is attempted to be applied. The
resulting tiller pod will be stuck in a Pending state.

The deployment strategy provided by 'helm init' is unspecified. This
change will allow one additional pod (current + new) and one unavailable
pod (current) during an update. The maxUnavailable setting allows the
tiller pod to be deleted which will release its ports, thus allowing the
patch deployment to spin up an new pod to a Running state.

This patching ensures that on an installed system where tiller has been
manually removed and re-applied via 'helm init', it is patched
appropriately.

Change-Id: I356545d05a585f7cbbbd5ca5071aa834fb086c31
Depends-On: https://review.opendev.org/#/c/725705/
Closes-Bug: #1876396
Signed-off-by: Robert Church <robert.church@windriver.com>