Add checks in host api for device image update
The host-device-image-update command is enhanced with a semantic
check to prevent updating device images on a disabled host.

The host-lock 'check_lock_worker' semantic check is enhanced to
prevent locking a host when device image update is in progress.

The host-swact 'check_swact' semantic check is enhanced to prevent
swacting when one of the controllers is in the process of updating
device images.

Story: 2006740
Task: 39498

Change-Id: I99fce8a34b1351fa868c5a45434f8a9d62a2a912
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>