Verify docker registry certificate with system trusted CAs
The issue was caused by sysinv verifying the registry certificate with
only itself. This works with the self-signed certificate installed by
the system but not if the user installs a new certificate.

This commit now checks the registry certificate against system trusted
CAs in addition to against itself. This ensures system registry
commands work with both the self-signed certificates and certificates
installed by the user that are signed by an actual CA.

Exception handling has also been improved so that sysinv can detect an
exception thrown while communicating with the registry. This is now
logged and returned to the user as opposed to hanging the sysinv
registry commands.

Closes-Bug: 1863624

Change-Id: I1170e296d5037716f2f800eede32e15861d038ce
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>