SR-IOV device plugin support for FPGA FEC devices
This commit introduces SR-IOV device plugin support for forward
error correction (FEC) devices that are enabled on an Intel
N3000 FPGA device.

The FEC device is intended to be used with a DPDK application,
with the device itself being bound to the igb_uio driver. As
such, support has been added for configuring the PF and VF
driver for the device, as well as the number of requested
VFs.

An example usage for setting the drivers and VFs is as follows:

system host-device-modify <host> <device_name> \
  --driver <driver> \
  --vf-driver <driver> \
  -N <number of vfs>

Story: 2006740
Task: 39947

Depends-On: https://review.opendev.org/733723
Change-Id: Iee0f0e449652219939ed5369423efdfb906ca3f1
Signed-off-by: Steven Webster <Steven.Webster@windriver.com>