Device image update alarm and audit
When the first device image is imported to the system, the alarm
"device image update is in progress" is raised. When all the
devices have been updated and rebooted, this alarm is cleared.

A periodic audit is added to check if any of the image updates
has timed out and declared it as failed.

Story: 2006740
Task: 39498

Change-Id: I44a027ee384cfa5b96d5b426b06173b54187fda9
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>