Monitor the owership of dir /home/sysadmin/.kube
This commit checks the owner of the /home/sysadmin/.kube directory.
The owner of this directory should be sysadmin: sys_protected for
Kubenetes work properly. This function is added into the
sysinv_conductor_monitor function in sysinv-conductor OCF script. The
.kube directory is created by the kubectl command, and the owership
may changed to root: root in subcloud controller  for initialize
helm before helm v3. Therefore this workaround should be deleted
afterwards

Test:
Deploy a DC system, check the .kube directory ownership in subcloud
controller after the controller is unlocked. The ownership should be
sysadmin: sys_protected.

Change-Id: I9be0cd558d4f1083ba2fa1f15b717f3de4eb64c7
Closes-Bug: 1882100
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>