Run kubeadm command with specific configuration file
sysinv uses the kubeadm command to generate kubeadm join command.
However, it is not passing the --kubeconfig parameter to this command,
so the command will search for the kubeconfig file. Normally it finds
this file in /etc/kubernetes/admin.conf, but if a user accidentally
(or intentionally) creates a config file at /root/.kube/config,
this file can take precedence and cause the kubeadm command to fail.

This commit updated the command with /etc/kubernetes/admin.conf as its
configuration file explicitly.

Change-Id: I4fb6d6325ca2cd8c7ab28c0acb0efdf6b1fea45b
Closes-Bug: 1882678
Signed-off-by: Andy Ning <andy.ning@windriver.com>