Fix subcloud status migration script
If a system controller upgrade is performed when there is at least one
subcloud that has never been managed, data migration will fail due to
type mismatch. This is because the empty updated_at field of the
patching status record that is used to create the load status
record for the subcloud results in a "None" value used for the sql insert.
This commit resolves the issue.

Closes-Bug: 1883458
Change-Id: Idbe636c26f66f00725b48f935fac0a1d1cf44a89
Signed-off-by: Tee Ngo <tee.ngo@windriver.com>