Update conditions for config change and manifest apply
The runtime manifest apply for an oam config change was being
triggered on host-swact to the target controller after startup.
Thus, the config runtime manifest was being triggered even when
there was not an oam config change.

Update the runtime manifest apply for oam config to be triggered
on active controller startup after an oam configuration change.

During upgrades, disallow oam network changes as the configuration
affects the platform and kubernetes components dependent on the
OAM network.

Added https and admin endpoint changes to the audit as these can
set during configuration of a system.

Tests performed:
-Standard lab install with https. https was setup correctly.
-AIO-SX lab install with https lab AIO-SX. https was setup correctly.
Tested oam-modify required host-lock/unlock to clear the config out of
date alarm.
-DC lab install. Confirmed the admin endpoints are set to https.
-Standard lab install with https.  Tested  host-swact after initial
install and reinstall.
-oam-modify lock/unlock standby/host-swact and verify oam access.

Closes-Bug: 1874136
Co-authored-by: John Kung <john.kung@windriver.com>
Change-Id: Icec24c3095e5ab13454fe34c9b94da8f8e3eaa23
Signed-off-by: Kristine Bujold <kristine.bujold@windriver.com>