Fix data-migration failure when upgrading N+1 node
When upgrading to N+1 node, cluster IP address from admin.conf
needs to be replaced with the floating management IP address.
This is to prevent failure in data migration when kubernetes api
accesses dc-adminep-root-ca-certificate.
The update of cluster IP address is temporary and done on a
copy of the original /etc/kubernetes/admin.conf.
This commit applies changes from the N side.

Closes-Bug: 1883763

Change-Id: Ibc2dbbab5016d9c58248ebb5b0fb52299e99705c
Signed-off-by: Carmen Rata <carmen.rata@windriver.com>