Provide app rollback support over upgrades for non-decoupled apps
In STX 4.0, helm plugins were removed and delivered as part of the
application tarball. During platform upgrade, platform applications are
updated (uploaded and applied) to the new delivered versions. In the
case of an apply failure the application is rolled back to restore
application functionality.

The current decoupled app framework relies on the existence of a plugin
directory to signify that it is a system knowledgeable application. The
prior not decoupled applications, do not have this structure. This
update will identify them so their saved overrides can be used during
rollback.

Change-Id: Ie11d906c933d9fca8b6f698944d8c5b1544e4eba
Story: 2007403
Task: 40118
Signed-off-by: Robert Church <robert.church@windriver.com>