Add pre-checks for k8s app upgrades to validate app states
During an upgrade on an AIO-SX, images are automatically recovered after
migrating to the new release and the host is unlocked. While this is
occurring, applications may be in an 'applying' or 'restore-requested'
state. Prevent application updates until these states clear to provide
proper conditions for application rollback on failed upgrades.

This will fail the upgrade-activation step and log an appropriate
message.

Change-Id: I94af4cd79f0e792bcb05eef19999adfa5d166493
Story: 2007403
Task: 40118
Signed-off-by: Robert Church <robert.church@windriver.com>