Handle N3000 FPGA reset
A recent commit (6d639fca5d) was introduced that resets any Intel
N3000 FPGA devices on startup.  After an FPGA reset, it is possible
that the PCI addresses of the underlying FEC device may change, in
addition to the resetting of any configured drivers and SR-IOV VF
configuration on the FPGA FEC and NIC devices.

To handle the reset, it is required to do the following:

- Do not allow a user to configure an FEC device before the first
unlock (to allow the reset to occur).

- Delay the configuration of FEC and NIC devices until the reset
  has occurred.

Story: 2006740
Task: 39947

Change-Id: I6ef8225cce5bf2d4f0ded2e20b6375504a05e587
Signed-off-by: Steven Webster <steven.webster@windriver.com>