Provide plugin access beyond application removal
Application plugins are currently deactivated once the application is
removed and is no longer active in the k8s cluster. For more complex
applications, allow the plugin deactivation to be deferred so that the
plugins are still available for post removal platform configuration
actions.

After a platform configuration action only evaluate app re-applies for
'applied' applications.

Makes sure an app is active and applied before re-evaluating the
overrides when determining if a re-apply is required.

Remove log pollution from the HelmOperator. The logs aren't overly
useful. Move them from info to debug.

Add a synchronized lock around the key plugin setup/access methods in
the HelmOperator.

Change-Id: I2ba2f3f1da83aae73100507bd6795e68438940d3
Closes-Bug: #1884408
Signed-off-by: Robert Church <robert.church@windriver.com>