Implement a short wait before removing load files
Load import could fail in slow vbox or hardware if the copying
of the large iso file by sysinv api proxy takes longer than
the time sysinv conductor imports the load.

In this commit, the load import thread in sysinv conductor waits
30 seconds before removing the load files from the staging directory.
This gives sysinv api proxy ample amount of time to copy the load
files to dc-vault. The wait does not apply to requests not going
through the proxy. Sysinv api proxy will be refactored in a future
commit to enable copying of load files directly from the import
request.

Closes-Bug: 1884694
Change-Id: I7ba4af8d32707d3c0977b29c93ab89894bce70cd
Signed-off-by: Tee Ngo <Tee.Ngo@windriver.com>