Data migration code to set the deleted field to 0
The field does not have a default value.
This caused a subcloud cannot be unmanaged after an upgrade.

Test in a simulated environment:
Reproduced the issue with manually setting deleted filed to Null.
Deleted the subcloud_status record with endpoint dc-cert to
simulate pre-upgrade env,
run the sql script to simulate upgrade data migration.
Was able to successfully unmanage the subcloud after.

Change-Id: I4195c1901345c27d10b9dea59ee3f7acb9b1cd12
Story: 2007347
Task: 40113
Signed-off-by: Bin Qian <bin.qian@windriver.com>