Fix run_docker_login DNS lookup to properly handle IPv6
It turns out that the nslookup tool doesn't handle IPv6 lookups by
default, at least the way our DNS server is configured.

In order to deal with this we switch to using the "dig" command, which
allows asking for both the A and AAAA records (IPv4 and IPv6,
respectively) in a single command.  A return code of 0 means we got
a response from the server, and if we did get a response then a
non-empty output  means that we successfully looked up the hostname.

Change-Id: I6efec9fe20ba192422ae19989430b1e8866f70dd
Partial-Bug: 1886121
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>